refactor: User 엔티티 변경에 따른 Folder 관련 코드 수정
